win32_installer: Re-organize the MSVS property sheets
We can drop the vs80 and vs90 property sheets, as it has been a while since we
can build gtkmm with Visual Studio 2005/2008.  Also, update the 2010 one so
that things fit with the current situation, as format-wise it is very much the
same with the format of the Visual Studio 2015+ property sheets, so make sure
that we dist that file as well.

Also, stop setting things up for libglade/libglademm (dropped for gtkmm-3.x)
and libxml++ (gtkmm applications may not be using libxml++), and clean things
up in there as well.

For the Visual Studio builds, the file names are based on those used by the
Meson builds.